Fitness function 4 bug fix and change ff to return double, not float

+++ simsj/sim/drift.cpp
@@ -89,9 +89,9 @@ int main (int argc, char** argv)
     array<genosect_t, N_Genes> genome;
 
     unsigned int lastgen = 0;
-    float f = 0.0f;
+    double f = 0.0;
 #ifdef RECORD_ALL_FITNESS
-    float lastf = 0.0f;
+    double lastf = 0.0;
     AllBasins ab1;
 #endif
 
@@ -119,7 +119,7 @@ int main (int argc, char** argv)
 #endif
 
         // Test fitness to determine whether we should evolve.
-        if (f == 1.0f) {
+        if (f == 1.0) {
             DBG ("Fitness max. F=" << f);
             // Record data
             generations.push_back (gen-lastgen);
@@ -130,7 +130,7 @@ int main (int argc, char** argv)
                 vector<NetInfo> vni;
                 netinfo.push_back (vni);
             }
-            lastf = 0.0f;
+            lastf = 0.0;
 #endif
             // Reset loop variables
             need_new_genome = true;
